AnyTest
[]

+++ AssaultSrtProvider/Representation/Snapshot.cs
@@ -3,8 +3,8 @@ namespace AssaultSrtProvider.Representation
     public class Snapshot
     {
         public readonly Tag[] Tags;
-        public readonly int Start;
-        public readonly int End;
+        public readonly float Start;
+        public readonly float End;
         
         public Snapshot(Tag[] tags, int start, int end)
         {
